+++ src/main/java/school/faang/user_service/dto/user/GetUserRequest.java
@@ -1,6 +1,7 @@
 package school.faang.user_service.dto.user;
 
 import lombok.Data;
+import school.faang.user_service.dto.UserDto;
 
 @Data
 public class GetUserRequest {
