+++ src/main/java/school/faang/user_service/dto/user/UserFilter.java
@@ -1,5 +1,6 @@
 package school.faang.user_service.dto.user;
 
+import school.faang.user_service.dto.UserDto;
 import school.faang.user_service.entity.User;
 
 import java.util.List;
