+++ src/main/java/school/faang/user_service/dto/user/UserNameFilter.java
@@ -1,6 +1,7 @@
 package school.faang.user_service.dto.user;
 
 import org.springframework.stereotype.Component;
+import school.faang.user_service.dto.UserDto;
 import school.faang.user_service.entity.User;
 
 import java.util.List;
