+++ src/main/java/school/faang/user_service/service/UserService.java
@@ -1,8 +1,17 @@
 package school.faang.user_service.service;
 
 import org.springframework.http.ResponseEntity;
+import school.faang.user_service.dto.TariffDto;
 import school.faang.user_service.dto.UserDto;
+import school.faang.user_service.dto.user.GetUserRequest;
+
+import java.util.List;
 
 public interface UserService {
     ResponseEntity<UserDto> getUser(long userId);
+
+    TariffDto buyUserTariff(TariffDto tariffDto, Long userId);
+
+    List<UserDto> findUsersByFilter(GetUserRequest request);
+
 }
